Fix default inAudienceValidationFunction
The function had a statement that performed no value:
_.flatten(userRoles, this.roleToAudienceMapFunction) resulted in the
userRoles array being flattened, but it should have been mapping the
userRoles with the roleToAudienceMapFunction. Fixed this behavior and
also extended the behavior so the roleToAudienceMapFunction can return
an array of values, which will then be flattened into the final
userAudience list.